django_backend/requirements.txt
Django==5.1.4
pipenv==2024.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
django-bootstrap5
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
urllib3==2.2.3
django-clearcache==1.2.1
djangorestframework
djoser
djangorestframework-simplejwt